2.3.1

+++ Spring-CRUD-master/src/main/java/web/dao/UserDAO.java
@@ -6,8 +6,8 @@ import java.util.List;
 public interface UserDAO {
 
     List<User>getAllUsers ();
-    User getUserById(long id);
+    User getUserById(int id);
     void addUser(User user);
-    void removeUser(long id);
+    void removeUser(int id);
     void updateUser(User user);
 }
