+++ Spring-CRUD-master/src/main/java/web/service/UserServiceImpl.java
@@ -25,7 +25,7 @@ public class UserServiceImpl implements UserService {
     }
 
     @Override
-    public Object getUserById(long id) {
+    public Object getUserById(int id) {
         return userDAO.getUserById(id);
     }
 
@@ -35,7 +35,7 @@ public class UserServiceImpl implements UserService {
     }
 
     @Override
-    public void removeUser(long id) {
+    public void removeUser(int id) {
         userDAO.removeUser(id);
     }
